함수를 utils.py로 옮김
